sql-fundamentals/database-metadata/pyproject.toml
[project]
name = "database-metadata"
version = "0.1.0"
description = "Extractor de metadatos de bases de datos PostgreSQL, SQL Server y Oracle con generación automática de documentación"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Marco Villarreal" }
]
keywords = ["database", "metadata", "documentation", "postgresql", "sqlserver", "oracle"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Documentation",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "psycopg[binary]>=3.2.0",
    "pyodbc>=5.2.0",
    "oracledb>=2.0.0",
    "PyYAML>=6.0.0",
    "Jinja2>=3.1.0",
    "colorama>=0.4.6",
]

[project.scripts]
db-metadata-postgres = "main_postgres_metadata:main"
db-metadata-mssql = "main_mssql_metadata:main"
db-metadata-oracle = "main_oracle_metadata:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["services"]
